refactor: consolidate shared HTTP client/server generator pieces
Factor HTTP_REASON_PHRASES, the HttpGlobalError alias block and the HttpError
class source into protocols/http/shared.ts. Both protocol files still emit their
own copy — they are independent modules — but the generator now has one source
for each. Generated output is byte-identical (all snapshots unchanged).
