+++ src/codegen/generators/typescript/channels/protocols/http/common-types.ts
@@ -11,54 +11,19 @@ import {
   renderOAuth2Helpers,
   renderSecurityTypes
 } from './security';
+import {
+  HTTP_GLOBAL_ERROR_ALIAS,
+  HTTP_REASON_PHRASES,
+  renderHttpErrorClass
+} from './shared';
 
-/**
- * Standard HTTP reason phrases used to bake a message into each generated
- * `handleHttpError` case. Keyed by status code; deterministic and
- * collision-free across operations that declare the same code.
- */
-const HTTP_REASON_PHRASES: Record<number, string> = {
-  400: 'Bad Request',
-  401: 'Unauthorized',
-  402: 'Payment Required',
-  403: 'Forbidden',
-  404: 'Not Found',
-  405: 'Method Not Allowed',
-  406: 'Not Acceptable',
-  407: 'Proxy Authentication Required',
-  408: 'Request Timeout',
-  409: 'Conflict',
-  410: 'Gone',
-  411: 'Length Required',
-  412: 'Precondition Failed',
-  413: 'Payload Too Large',
-  414: 'URI Too Long',
-  415: 'Unsupported Media Type',
-  416: 'Range Not Satisfiable',
-  417: 'Expectation Failed',
-  418: "I'm a Teapot",
-  421: 'Misdirected Request',
-  422: 'Unprocessable Entity',
-  423: 'Locked',
-  424: 'Failed Dependency',
-  425: 'Too Early',
-  426: 'Upgrade Required',
-  428: 'Precondition Required',
-  429: 'Too Many Requests',
-  431: 'Request Header Fields Too Large',
-  451: 'Unavailable For Legal Reasons',
-  500: 'Internal Server Error',
-  501: 'Not Implemented',
-  502: 'Bad Gateway',
-  503: 'Service Unavailable',
-  504: 'Gateway Timeout',
-  505: 'HTTP Version Not Supported',
-  506: 'Variant Also Negotiates',
-  507: 'Insufficient Storage',
-  508: 'Loop Detected',
-  510: 'Not Extended',
-  511: 'Network Authentication Required'
-};
+const CLIENT_HTTP_ERROR_DESCRIPTION = `/**
+ * Error thrown for non-OK HTTP responses.
+ *
+ * Carries the HTTP \`status\`, \`statusText\`, and the parsed response \`body\`
+ * (when the error response had a JSON body). Thrown by \`handleHttpError\` and
+ * routed through the \`onError\` hook / retry logic unchanged.
+ */`;
 
 /**
  * Build the body of the generated `handleHttpError` function from the set of
@@ -154,16 +119,7 @@ const API_KEY_DEFAULTS = {
 // Common Types - Shared across all HTTP client functions
 // ============================================================================
 
-/**
- * The global \`Error\`, captured under a name a payload model cannot take.
- *
- * A document is free to declare a schema called \`Error\` (it is the
- * conventional name for one), and its generated model is imported into this
- * module, shadowing the global for the whole file. Every reference below goes
- * through these aliases so that is harmless.
- */
-const HttpGlobalError = globalThis.Error;
-type HttpGlobalError = InstanceType<typeof globalThis.Error>;
+${HTTP_GLOBAL_ERROR_ALIAS}
 
 /**
  * Standard HTTP response interface that wraps fetch-like responses
@@ -192,26 +148,7 @@ export interface HttpClientResponse<T> {
   rawData: Record<string, any>;
 }
 
-/**
- * Error thrown for non-OK HTTP responses.
- *
- * Carries the HTTP \`status\`, \`statusText\`, and the parsed response \`body\`
- * (when the error response had a JSON body). Thrown by \`handleHttpError\` and
- * routed through the \`onError\` hook / retry logic unchanged.
- */
-export class HttpError extends HttpGlobalError {
-  status: number;
-  statusText: string;
-  body?: unknown;
-
-  constructor(message: string, status: number, statusText: string, body?: unknown) {
-    super(message);
-    this.name = 'HttpError';
-    this.status = status;
-    this.statusText = statusText;
-    this.body = body;
-  }
-}
+${renderHttpErrorClass(CLIENT_HTTP_ERROR_DESCRIPTION)}
 
 /**
  * HTTP request parameters passed to the request hook
